+++ src/app/core/components/Modal/TransactionModal/BakeryTransactionModal.tsx
@@ -3,15 +3,15 @@ import {
   ModalContainer,
   ModalContent,
   ModalHeading,
-  ShareButtons,
   transactionIcons,
   TransactionValue,
 } from "../ModalUI";
-import { TTransactionData, TTransactionStatus } from "../../../context/TransactionsContext/TransactionsReducer";
+import { TTransactionStatus } from "@/app/core//context/TransactionsContext/TransactionsReducer";
 import {
   TokenLabelContainer,
   TokenLabelText,
 } from "@/app/bakery/components/Swap/SwapUI";
+import Button from "@/app/core/components/Button";
 import { BreadIcon } from "../../Icons/TokenIcons";
 import { ExplorerLink } from "../../ExplorerLink";
 import { useTransactions } from "@/app/core/context/TransactionsContext/TransactionsContext";
@@ -39,6 +39,62 @@ const modalAdviceText: {
   CONFIRMED: "You have successfully baked",
   REVERTED: "Transaction Reverted",
 };
+
+
+function ShareButtons({ bakeValue }: { bakeValue: string }) {
+  function makeText(platform: 'X' | 'Warpcast') {
+    return `I just baked some BREAD to help fund on-chain post-capitalism thanks to ${platform === 'X' ? '@breadchain_' : 'breadchain.xyz'}!
+
+I grew the bakery to ${bakeValue} BREAD! \u{1F35E} \u{1F35E} \u{1F35E}
+
+https://app.breadchain.xyz`
+  }
+
+  const xText = encodeURIComponent(makeText('X'))
+  const wText = encodeURIComponent(makeText('Warpcast'))
+
+  const shareOnX = () => {
+    window.open(`https://twitter.com/intent/tweet?text=${xText}`, '_blank')
+  }
+  const shareOnW = () => {
+    window.open(`https://warpcast.com/~/compose?text=${wText}`, '_blank')
+  }
+
+  return (
+    <div className="w-full flex flex-col gap-4">
+      <Button
+        fullWidth={true}
+        size="large"
+        variant="secondary"
+        onClick={shareOnX}
+      >
+        Share on <img
+          className="inline ml-1"
+          src="/x_logo_white.png"
+          width={20}
+          height={20}
+          alt="X logo"
+        />
+      </Button>
+      <Button
+        fullWidth={true}
+        size="large"
+        variant="secondary"
+        onClick={shareOnW}
+      >
+        Share on <img
+          className="inline ml-1"
+          src="/warpcast_logo_round.png"
+          width={20}
+          height={20}
+          alt="Warpcast logo"
+        />
+      </Button>
+
+    </div>
+  );
+}
+
 
 export function BakeryTransactionModal({
   modalState,
